댓글에 관한 코드

+++ src/main/java/com/prj2spring/controller/comment/CommentController.java
@@ -3,6 +3,7 @@ package com.prj2spring.controller.comment;
 import com.prj2spring.domain.comment.Comment;
 import com.prj2spring.service.comment.CommentService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
@@ -38,7 +39,29 @@ public class CommentController {
     }
 
     @DeleteMapping("remove")
-    public void remove(@RequestBody Comment comment) {
-        service.remove(comment);
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity remove(@RequestBody Comment comment,
+                                 Authentication authentication) {
+        if (service.hasAccess(comment, authentication)) {
+            service.remove(comment);
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
+
+    }
+
+
+    @PutMapping("edit")
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity edit(@RequestBody Comment comment,
+                               Authentication authentication) {
+        if (service.hasAccess(comment, authentication)) {
+            service.update(comment);
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
     }
 }
